app/src/requirements.txt
azure-cognitiveservices-speech
ffmpeg
gunicorn
moviepy
nltk
pandas
pytest
requests
setuptools